chore: demo extract
