+++ src/client/theme-api/DumiDemo/index.tsx
@@ -1,6 +1,10 @@
-import React, { Suspense, type FC } from 'react';
+import { SP_ROUTE_PREFIX } from '@/constants';
+import { useAppData, useDemoData, useSiteData } from 'dumi';
+import React, { createElement, type FC } from 'react';
 import type { IPreviewerProps } from '../types';
-import DumiDemoBlock from './DemoBlock';
+
+import Previewer from 'dumi/theme/builtins/Previewer';
+import DemoErrorBoundary from './DemoErrorBoundary';
 
 export interface IDumiDemoProps {
   demo: {
@@ -10,11 +14,48 @@ export interface IDumiDemoProps {
   previewerProps: Omit<IPreviewerProps, 'asset' | 'children'>;
 }
 
-const InternalDumiDemo = (props: IDumiDemoProps) => (
-  <Suspense>
-    <DumiDemoBlock {...props} />
-  </Suspense>
-);
+const InternalDumiDemo = (props: IDumiDemoProps) => {
+  const { historyType } = useSiteData();
+  const { basename } = useAppData();
+  const demoInfo = useDemoData(props.demo.id);
+
+  // hide debug demo in production
+  if (process.env.NODE_ENV === 'production' && props.previewerProps.debug)
+    return null;
+
+  if (!demoInfo) {
+    return `Demo '${props.demo.id}' not found!`;
+  }
+
+  const { component, asset } = demoInfo;
+
+  const demoNode = (
+    <DemoErrorBoundary>{createElement(component)}</DemoErrorBoundary>
+  );
+
+  if (props.demo.inline) {
+    return demoNode;
+  }
+
+  const isHashRoute = historyType === 'hash';
+
+  return (
+    <Previewer
+      asset={asset}
+      demoUrl={
+        // allow user override demoUrl by frontmatter
+        props.previewerProps.demoUrl ||
+        // when use hash route, browser can automatically handle relative paths starting with #
+        `${isHashRoute ? `#` : ''}${basename}${SP_ROUTE_PREFIX}demos/${
+          props.demo.id
+        }`
+      }
+      {...props.previewerProps}
+    >
+      {props.previewerProps.iframe ? null : demoNode}
+    </Previewer>
+  );
+};
 
 export const DumiDemo: FC<IDumiDemoProps> = React.memo(
   InternalDumiDemo,
